reorganized functions, reorganized nbs, first look at sb1
